+++ decidim-admin/app/helpers/decidim/admin/settings_helper.rb
@@ -45,6 +45,8 @@ module Decidim
             form.send(:translated, form_method, name, options)
           elsif form_method == :collection_radio_buttons
             render_enum_form_field(form, attribute, name, i18n_scope, options)
+          elsif attribute.type.to_sym == :scope
+            scopes_picker_field(form, name)
           else
             form.send(form_method, name, options)
           end