+++ decidim-assemblies/app/models/decidim/assembly.rb
@@ -24,7 +24,7 @@ module Decidim
     include Decidim::HasAttachmentCollections
     include Decidim::Participable
     include Decidim::Publicable
-    include Decidim::Scopable
+    include Decidim::ScopableParticipatorySpace
     include Decidim::Followable
     include Decidim::HasReference
     include Decidim::Traceable