+++ decidim-core/app/models/decidim/action_log.rb
@@ -5,7 +5,7 @@ module Decidim
   # for transparency reasons, to log all actions so all other users can
   # see the actions being performed.
   class ActionLog < ApplicationRecord
-    include Decidim::Scopable
+    include Decidim::ScopableParticipatorySpace
 
     belongs_to :organization,
                foreign_key: :decidim_organization_id,