+++ decidim-proposals/app/forms/decidim/proposals/admin/proposal_form.rb
@@ -28,9 +28,8 @@ module Decidim
         validates :address, geocoding: true, if: -> { current_component.settings.geocoding_enabled? }
         validates :category, presence: true, if: ->(form) { form.category_id.present? }
         validates :scope, presence: true, if: ->(form) { form.scope_id.present? }
+        validates :scope_id, scope_belongs_to_component: true, if: ->(form) { form.scope_id.present? }
         validates :meeting_as_author, presence: true, if: ->(form) { form.created_in_meeting? }
-
-        validate :scope_belongs_to_participatory_space_scope
 
         validate :notify_missing_attachment_if_errored
 
@@ -54,11 +53,11 @@ module Decidim
           @category ||= categories.find_by(id: category_id)
         end
 
-        # Finds the Scope from the given decidim_scope_id, uses participatory space scope if missing.
+        # Finds the Scope from the given scope_id, uses participatory space scope if missing.
         #
         # Returns a Decidim::Scope
         def scope
-          @scope ||= @scope_id ? current_participatory_space.scopes.find_by(id: @scope_id) : current_participatory_space.scope
+          @scope ||= @scope_id ? current_component.scopes.find_by(id: @scope_id) : current_component.scope
         end
 
         # Scope identifier
@@ -107,10 +106,6 @@ module Decidim
         end
 
         private
-
-        def scope_belongs_to_participatory_space_scope
-          errors.add(:scope_id, :invalid) if current_participatory_space.out_of_scope?(scope)
-        end
 
         # This method will add an error to the `attachment` field only if there's
         # any error in any other field. This is needed because when the form has